Terrain vertices coloured by normalized height.

+++ Project/Time.cpp
@@ -1,12 +1,12 @@
 #include "Time.h"
 
-float Time::prevTime = 0;
-float Time::currentTime = 0;
+double Time::prevTime = 0;
+double Time::currentTime = 0;
 float Time::deltaTime = 0;
 
 void Time::update()
 {
 	prevTime = currentTime;
 	currentTime = glfwGetTime();
- 	deltaTime = float(currentTime - prevTime);
+ 	deltaTime = currentTime - prevTime;
 }
